requirements.txt
Flask
opencv-python-headless
numpy
pywavelets
Pillow
cryptography